Update version number, dependencies, and projector metadata
This update increments the library version to 1.7.1, adjusts the version of certain dependencies, and modifies the project metadata including the list of authors, contact information, and relevant links. It further includes changes in the internal code for optimisations and refactors the worker threads' setup to respect system memory availability with a new boolean `allow_swap`.