Refine description of AIS messages in readme.rst
This commit refines the explanation of Automatic Identification System (AIS) messages in the readme.rst file. It elaborates the types of information contained in AIS messages and highlights the limitation of open-source AIS data, primarily being positional. The methods and challenges of AIS message collection and encoding are also expanded upon.