Update bathymetry.py for improved file handling and readability
This update refactors code in bathymetry.py for better efficiency and readability. The changes include improved handling of raster files with a more convenient way to download and extract archives. The code is simplified in some areas for better readability as well. Comments throughout the code are updated for clarity, and unnecessary import statements and variables are removed. The filename extension and download URL have also been corrected. Relevant testing workflows in CI.yml have been uncommented to embrace these changes.