Remove unnecessary variable type conversions in db.rs
The conversion of variables to an integer format using "as i32" has been removed in several instances in the db.rs file. The variables mmsi and imo_number were originally being unnecessarily converted to an integer type, but this is not actually required within the scope of the functions they are used in.